Fix on tournament results

+++ src/main/java/org/example/model/Rank.java
@@ -7,11 +7,12 @@ import java.util.Objects;
 public class Rank implements Comparable<Rank>, Serializable {
     public float points;
     public String name;
+
     public Rank(String name){
         this.name = name;
         this.points = 0;
     }
-    public Rank(String name, int points){
+    public Rank(String name, float points){
         this.name = name;
         this.points = points;
     }
